refactor(telemetry): Move shared methods to base class.
Add tests/unit/test_unit_telemetry.
