+++ src/any_agent/evaluation/evaluate.py
@@ -10,7 +10,7 @@ from any_agent.evaluation.evaluators import (
 from any_agent.evaluation.results_saver import save_evaluation_results
 from any_agent.evaluation.test_case import TestCase
 from any_agent.logging import logger
-from any_agent.telemetry.telemetry import TelemetryProcessor
+from any_agent.telemetry.base import TelemetryProcessor
 
 
 def evaluate_telemetry(test_case: TestCase, telemetry_path: str) -> bool:
