+++ src/any_agent/telemetry/smolagents_telemetry.py
@@ -174,27 +174,3 @@ class SmolagentsTelemetryProcessor(TelemetryProcessor):
             agent_info["token_usage"] = token_counts
 
         return agent_info
-
-    def _extract_telemetry_data(self, telemetry: list[dict[str, Any]]) -> list[dict]:
-        """Extract LLM calls and tool calls from SmoL Agents telemetry."""
-        calls = []
-
-        for span in telemetry:
-            calls.append(self.extract_interaction(span)[1])
-        return calls
-
-    def extract_interaction(self, span: dict[str, Any]) -> tuple[str, dict[str, Any]]:
-        """Extract interaction details from a span."""
-        attributes = span.get("attributes", {})
-        span_kind = attributes.get("openinference.span.kind", "")
-
-        if span_kind == "LLM" or "LiteLLMModel.__call__" in span.get("name", ""):
-            return "LLM", self._extract_llm_interaction(span)
-        if span_kind == "CHAIN":
-            return "CHAIN", self._extract_chain_interaction(span)
-        if span_kind == "AGENT":
-            return "AGENT", self._extract_agent_interaction(span)
-        if "tool.name" in attributes or span.get("name", "").startswith("SimpleTool"):
-            return "TOOL", self._extract_tool_interaction(span)
-        msg = f"Unknown span kind: {span_kind} or unsupported span name: {span.get('name', '')}"
-        raise ValueError(msg)
